pyproject.toml
[tool.poetry]
name = "piqueserver-toolsapi"
version = "0.1.0"
description = ""
authors = ["NotAFile <notafile@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.7"
pillow = "^6.2.1"
flask = "^1.1.1"

[tool.poetry.dev-dependencies]
pytest = "^5.2"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
